Provide an interface for grpc client therefore allowing the impl to be final while being able to mock the interface.
